requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=0.0.8
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Web application dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0

# GCP dependencies
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
redis[hiredis]>=5.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
